// src/app/models/asset-acquisition-type-audit.model.ts
export class AssetAcquisitionTypeAudit {
    AssetAcquisitionTypeAuditId: number;
    id: number;
    identification: string;
    name: string;
    memo: string;
    createdBy: string;
    updatedBy: string;
    createdDate: Date;
    updatedDate: Date;
    isDeleted: boolean;
    isActive: boolean;
}